+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ aiosmtplib changes
 
 - Bugfix: Badly encoded server response messages are now decoded to utf-8,
   with error chars escaped.
+- Cleanup: Removed handling for exceptions not raised by asyncio (in
+  SMTPProtocol._readline)
 
 
 1.0.3